requirements.txt
glob3
re
json
fitz
itemgetter
datetime
camelot-py[cv]
warnings
